feat: launch all gpio nodes
